chore: refactor core folder structure - module based
