+++ core/src/browser/extension.ts
