+++ core/src/browser/extensions/index.ts
