+++ core/src/browser/index.ts
@@ -1,0 +1,29 @@
+/**
+ * Export Core module
+ * @module
+ */
+export * from './core'
+
+/**
+ * Export Event module.
+ * @module
+ */
+export * from './events'
+
+/**
+ * Export Filesystem module.
+ * @module
+ */
+export * from './fs'
+
+/**
+ * Export Extension module.
+ * @module
+ */
+export * from './extension'
+
+/**
+ * Export all base extensions.
+ * @module
+ */
+export * from './extensions'
