+++ core/src/index.ts
@@ -2,42 +2,13 @@
  * Export all types.
  * @module
  */
-export * from './types/index'
+export * from './types'
 
 /**
- * Export all routes
- */
-export * from './api'
-
-/**
- * Export Core module
+ * Export browser module
  * @module
  */
-export * from './core'
-
-/**
- * Export Event module.
- * @module
- */
-export * from './events'
-
-/**
- * Export Filesystem module.
- * @module
- */
-export * from './fs'
-
-/**
- * Export Extension module.
- * @module
- */
-export * from './extension'
-
-/**
- * Export all base extensions.
- * @module
- */
-export * from './extensions/index'
+export * from './browser'
 
 /**
  * Declare global object
