+++ core/src/node/index.ts
@@ -5,4 +5,4 @@ export * from './extension/store'
 export * from './api'
 export * from './helper'
 export * from './../types'
-export * from './../api'
+export * from '../types/api'
