+++ core/src/types/index.ts
@@ -8,3 +8,4 @@ export * from './file'
 export * from './config'
 export * from './huggingface'
 export * from './miscellaneous'
+export * from './api'
